# api/v1/openapi.yaml
openapi: 3.0.3
info:
  title: Airport Flight Data Collector API
  description: |
    Comprehensive REST and GraphQL API for accessing flight data, predictions, and analytics.
    Features include real-time flight status, ML-based delay predictions, pattern detection,
    data export, webhook integrations, and advanced analytics for SFO, YYZ, and YVR airports.
  version: 1.0.0
  contact:
    name: API Support
    url: https://github.com/czhaoca/airport-flight-data-collector/issues
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001/api/v1
    description: Development server
  - url: https://api.flightdata.example.com/v1
    description: Production server
security:
  - BearerAuth: []
  - ApiKeyAuth: []
...
paths:
...
  /flights:
    get:
      tags:
        - Flights
      summary: Get flight data
      description: Retrieve flight data with various filters
      parameters:
        - $ref: '#/components/parameters/Airport'
        - $ref: '#/components/parameters/FlightType'
        - $ref: '#/components/parameters/StartDate'
        - $ref: '#/components/parameters/EndDate'
        - $ref: '#/components/parameters/Status'
        - $ref: '#/components/parameters/Airline'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlightListResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '429':
          $ref: '#/components/responses/RateLimitError'
  /flights/{id}:
    get:
      tags:
        - Flights
      summary: Get specific flight
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Flight ID
      responses:
        '200':
          description: Flight found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlightDetailResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
  /predictions/delay:
    post:
      tags:
        - Predictions
      summary: Predict delay for a flight
      description: Get ML-based delay prediction for a specific flight
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DelayPredictionRequest'
      responses:
        '200':
          description: Delay prediction generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DelayPredictionResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /predictions/batch:
    post:
      tags:
        - Predictions
      summary: Batch delay predictions
      description: Get delay predictions for multiple flights at once
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchPredictionRequest'
      responses:
        '200':
          description: Batch predictions generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchPredictionResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /predictions/upcoming:
    get:
      tags:
        - Predictions
      summary: Get predictions for upcoming flights
      description: Retrieve delay predictions for flights in the next few hours
      parameters:
        - $ref: '#/components/parameters/Airport'
        - name: hours
          in: query
          description: Number of hours ahead to look
          schema:
            type: integer
            minimum: 1
            maximum: 24
            default: 6
        - name: minRisk
          in: query
          description: Minimum risk level to include
          schema:
            type: string
            enum: [low, medium, high]
      responses:
        '200':
          description: Upcoming flight predictions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpcomingPredictionsResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /predictions/high-risk:
    get:
      tags:
        - Predictions
      summary: Get high-risk flight alerts
      description: Retrieve flights with high delay probability
      parameters:
        - $ref: '#/components/parameters/Airport'
        - name: threshold
          in: query
          description: Minimum delay probability threshold
          schema:
            type: number
            minimum: 0.5
            maximum: 1.0
            default: 0.8
      responses:
        '200':
          description: High-risk flight alerts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HighRiskAlertsResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /export/flights:
    get:
      tags:
        - Export
      summary: Export flight data
      description: Export flight data in various formats
      parameters:
        - $ref: '#/components/parameters/Airport'
        - $ref: '#/components/parameters/StartDate'
        - $ref: '#/components/parameters/EndDate'
        - name: format
          in: query
          description: Export format
          schema:
            type: string
            enum: [json, csv, parquet]
            default: json
        - name: fields
          in: query
          description: Comma-separated list of fields to include
          schema:
            type: string
            example: 'flightNumber,airline,status,delay'
      responses:
        '200':
          description: Flight data export
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportResponse'
            text/csv:
              schema:
                type: string
            application/octet-stream:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /batch:
    post:
      tags:
        - Batch
      summary: Execute batch operations
      description: Perform multiple operations in a single request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchRequest'
      responses:
        '200':
          description: Batch operations completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    Airport:
      name: airport
      in: query
      description: Filter by airport code
      schema:
        type: string
        enum: [SFO, YYZ, YVR]
    FlightType:
      name: type
      in: query
      description: Filter by flight type
      schema:
        type: string
        enum: [arrival, departure]
    StartDate:
      name: startDate
      in: query
      description: Start date (inclusive)
      schema:
        type: string
        format: date
        example: '2025-01-01'
    EndDate:
      name: endDate
      in: query
      description: End date (inclusive)
      schema:
        type: string
        format: date
        example: '2025-01-31'
    Status:
      name: status
      in: query
      description: Filter by flight status
      schema:
        type: string
        enum: [ON_TIME, DELAYED, CANCELLED, DEPARTED, ARRIVED, BOARDING]
    Airline:
      name: airline
      in: query
      description: Filter by airline code
      schema:
        type: string
        example: 'UA'
    Limit:
      name: limit
      in: query
      description: Maximum number of results
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 100
    Offset:
      name: offset
      in: query
      description: Number of results to skip
      schema:
        type: integer
        minimum: 0
        default: 0
    Sort:
      name: sort
      in: query
      description: Sort order
      schema:
        type: string
        enum: [scheduledTime, actualTime, flightNumber, status]
        default: scheduledTime
...
  schemas:
...
    Flight:
      type: object
      properties:
        id:
          type: string
          description: Unique flight identifier
        flightNumber:
          type: string
          example: 'UA123'
        airline:
          $ref: '#/components/schemas/Airline'
        origin:
          type: string
          example: 'LAX'
        destination:
          type: string
          example: 'SFO'
        scheduledTime:
          type: string
          format: date-time
        actualTime:
          type: string
          format: date-time
          nullable: true
        estimatedTime:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
          enum: [ON_TIME, DELAYED, CANCELLED, DEPARTED, ARRIVED, BOARDING]
        gate:
          type: string
          nullable: true
        terminal:
          type: string
          nullable: true
        type:
          type: string
          enum: [arrival, departure]
        aircraft:
          type: string
          nullable: true
        delay:
          type: integer
          description: Delay in minutes
          nullable: true
    Airline:
      type: object
      properties:
        code:
          type: string
          example: 'UA'
        name:
          type: string
          example: 'United Airlines'
...
    FlightListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Flight'
        pagination:
          $ref: '#/components/schemas/Pagination'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    FlightDetailResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Flight'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...
    Pagination:
      type: object
      properties:
        limit:
          type: integer
        offset:
          type: integer
        total:
          type: integer
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
    ResponseMeta:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
        cached:
          type: boolean
          default: false
    DelayPredictionRequest:
      type: object
      required:
        - scheduledTime
        - airline
        - origin
      properties:
        flightNumber:
          type: string
          example: 'UA123'
        airline:
          type: string
          example: 'UA'
        origin:
          type: string
          example: 'SFO'
        destination:
          type: string
          example: 'LAX'
        scheduledTime:
          type: string
          format: date-time
          example: '2025-08-02T08:00:00Z'
        aircraft:
          type: string
          example: 'B737'
        gate:
          type: string
          example: 'A12'
    DelayPredictionResponse:
      type: object
      properties:
        success:
          type: boolean
        prediction:
          type: object
          properties:
            delayProbability:
              type: number
              format: float
              minimum: 0
              maximum: 1
              description: Probability of delay (0-1)
            expectedDelayMinutes:
              type: integer
              minimum: 0
              description: Expected delay in minutes
            riskLevel:
              type: string
              enum: [low, medium, high]
            confidence:
              type: number
              format: float
              minimum: 0
              maximum: 1
            factors:
              type: array
              items:
                type: object
                properties:
                  factor:
                    type: string
                  impact:
                    type: number
                    format: float
                  description:
                    type: string
    BatchPredictionRequest:
      type: object
      required:
        - flights
      properties:
        flights:
          type: array
          items:
            $ref: '#/components/schemas/DelayPredictionRequest'
          maxItems: 100
    BatchPredictionResponse:
      type: object
      properties:
        success:
          type: boolean
        predictions:
          type: array
          items:
            type: object
            properties:
              flightId:
                type: string
              prediction:
                $ref: '#/components/schemas/DelayPredictionResponse/properties/prediction'
              error:
                type: string
                nullable: true
    UpcomingPredictionsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            type: object
            properties:
              flight:
                $ref: '#/components/schemas/Flight'
              prediction:
                $ref: '#/components/schemas/DelayPredictionResponse/properties/prediction'
        meta:
          type: object
          properties:
            totalFlights:
              type: integer
            timeRange:
              type: string
            generatedAt:
              type: string
              format: date-time
    HighRiskAlertsResponse:
      type: object
      properties:
        success:
          type: boolean
        alerts:
          type: array
          items:
            type: object
            properties:
              flight:
                $ref: '#/components/schemas/Flight'
              prediction:
                $ref: '#/components/schemas/DelayPredictionResponse/properties/prediction'
              alertLevel:
                type: string
                enum: [warning, critical]
              message:
                type: string
...
    ExportResponse:
      type: object
      properties:
        success:
          type: boolean
        exportId:
          type: string
        downloadUrl:
          type: string
        format:
          type: string
        recordCount:
          type: integer
        fileSize:
          type: integer
          description: File size in bytes
        expiresAt:
          type: string
          format: date-time
...
    BatchRequest:
      type: object
      required:
        - operations
      properties:
        operations:
          type: array
          items:
            type: object
            required:
              - operation
              - path
            properties:
              operation:
                type: string
                enum: [GET, POST, PUT, DELETE]
              path:
                type: string
                example: '/flights?airport=SFO&limit=10'
              body:
                type: object
                nullable: true
              id:
                type: string
                description: Optional ID to correlate request with response
          maxItems: 100
    BatchResponse:
      type: object
      properties:
        success:
          type: boolean
        results:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                nullable: true
              status:
                type: integer
              response:
                type: object
              error:
                type: string
                nullable: true
        summary:
          type: object
          properties:
            totalOperations:
              type: integer
            successfulOperations:
              type: integer
            failedOperations:
              type: integer
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
  responses:
    BadRequestError:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: BAD_REQUEST
            message: Invalid date format
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: UNAUTHORIZED
            message: Invalid or missing authentication token
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: NOT_FOUND
            message: Flight not found
    RateLimitError:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          description: Request limit per window
          schema:
            type: integer
        X-RateLimit-Remaining:
          description: Remaining requests in window
          schema:
            type: integer
        X-RateLimit-Reset:
          description: Time when rate limit resets
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: RATE_LIMIT_EXCEEDED
            message: Too many requests. Please try again later.